rename top bar message
